refactor: move `FolderSync` to `storage`
