+++ app/src/main/java/com/osfans/trime/ui/main/PrefMainActivity.kt
@@ -34,8 +34,8 @@ import com.osfans.trime.core.RimeLifecycle
 import com.osfans.trime.daemon.RimeDaemon
 import com.osfans.trime.data.AppPrefs
 import com.osfans.trime.data.sound.SoundEffectManager
+import com.osfans.trime.data.storage.FolderSync
 import com.osfans.trime.databinding.ActivityPrefBinding
-import com.osfans.trime.ime.core.OneWayFolderSync
 import com.osfans.trime.ui.setup.SetupActivity
 import com.osfans.trime.util.progressBarDialogIndeterminate
 import com.osfans.trime.util.rimeActionWithResultDialog
@@ -189,10 +189,10 @@ class PrefMainActivity : AppCompatActivity() {
         val userDirUri = AppPrefs.defaultInstance().profile.userDataDir
         val shareDirUri = AppPrefs.defaultInstance().profile.sharedDataDir
 
-        OneWayFolderSync(this, userDirUri).copyAll((AppPrefs.Profile.getAppUserDir()))
+        FolderSync(this, userDirUri).copyAll((AppPrefs.Profile.getAppUserDir()))
 
         if (shareDirUri != userDirUri && shareDirUri.isNotBlank()) {
-            OneWayFolderSync(this, shareDirUri).copyAll((AppPrefs.Profile.getAppShareDir()))
+            FolderSync(this, shareDirUri).copyAll((AppPrefs.Profile.getAppShareDir()))
         }
     }
